+++ src/fs_trackdraw_qt.py
@@ -2,85 +2,13 @@ import os
 import math
 import csv
 import numpy as np
-from PyQt5.QtWidgets import (QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QLabel, 
+from PyQt5.QtWidgets import (QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QLabel, 
                             QPushButton, QLineEdit, QFileDialog, QMessageBox)
 from PyQt5.QtCore import Qt, QPointF
-from PyQt5.QtGui import QImage, QPixmap, QPainter, QPen, QColor, QPolygonF
-from PIL import Image
-import cv2
 import yaml
-from shapely.geometry import Polygon, LineString
 from scipy.interpolate import splprep, splev
-
-def create_closed_spline(control_points, num_points=100):
-    """
-    Compute a smooth, closed B-spline from a list of control points.
-    """
-    pts = np.array(control_points)
-    # add first point to end to close the loop
-    pts = np.vstack((pts, pts[0]))
-    # if we have only one point, return it
-    if len(pts) <= 2:
-        return pts
-    tck, _ = splprep([pts[:, 0], pts[:, 1]], s=0, per=True)
-    u_fine = np.linspace(0, 1, num_points)
-    x_fine, y_fine = splev(u_fine, tck)
-    return np.column_stack((x_fine, y_fine))
-
-
-def robust_parallel_offset(ls, distance, side, join_style=2):
-    """
-    Compute a parallel offset of a LineString using shapely's parallel_offset.
-    If the result is a MultiLineString, return the longest LineString.
-    """
-    offset = ls.parallel_offset(distance, side, join_style=join_style)
-    if offset.is_empty:
-        return None
-    if offset.geom_type == "MultiLineString":
-        lines = list(offset)
-        lines.sort(key=lambda l: l.length, reverse=True)
-        return lines[0]
-    return offset
-
-
-def generate_offset_boundaries(track_points, track_width_meters, px_per_m):
-    """
-    Compute left and right boundaries as parallel offsets from the centerline.
-    """
-    track_width_px = track_width_meters * px_per_m
-    ls = LineString(track_points)
-    left_offset = robust_parallel_offset(ls, track_width_px / 2.0, 'left', join_style=2)
-    right_offset = robust_parallel_offset(ls, track_width_px / 2.0, 'right', join_style=2)
-    if left_offset is None or right_offset is None:
-        return None, None
-    left_coords = list(left_offset.coords)
-    right_coords = list(right_offset.coords)
-    return np.array(left_coords), np.array(right_coords)
-
-
-def sample_cones(boundary, cone_spacing_meters, px_per_m):
-    """Sample points along a boundary so that they are approximately cone_spacing_meters apart."""
-    cone_spacing_px = cone_spacing_meters * px_per_m
-    pts = boundary
-    if len(pts) < 2:
-        return pts
-    distances = [0]
-    for i in range(1, len(pts)):
-        d = math.hypot(pts[i][0] - pts[i-1][0], pts[i][1] - pts[i-1][1])
-        distances.append(distances[-1] + d)
-    total_length = distances[-1]
-    num_cones = max(2, int(total_length // cone_spacing_px))
-    sample_d = np.linspace(0, total_length, num_cones)
-    sampled = []
-    for sd in sample_d:
-        for i in range(1, len(distances)):
-            if distances[i] >= sd:
-                t = (sd - distances[i-1]) / (distances[i] - distances[i-1])
-                x = pts[i-1][0] + t * (pts[i][0] - pts[i-1][0])
-                y = pts[i-1][1] + t * (pts[i][1] - pts[i-1][1])
-                sampled.append((x, y))
-                break
-    return np.array(sampled)
+from utils_qt import (create_closed_spline, generate_offset_boundaries, sample_cones)
+from track_canvas_qt import TrackCanvas
 
 
 class FSTrackDraw(QMainWindow):
@@ -109,29 +37,6 @@ class FSTrackDraw(QMainWindow):
             occup_img_file = config.get('occ_img_path', '')
             self.fpath_location_sat_img = os.path.join(self.folderpath_location, sat_img_file)
             self.fpath_location_occup_img = os.path.join(self.folderpath_location, occup_img_file)
-
-        # Load images
-        self.sat_image = Image.open(self.fpath_location_sat_img)
-        self.sat_qimage = QImage(self.fpath_location_sat_img)
-        self.sat_pixmap = QPixmap.fromImage(self.sat_qimage)
-        
-        # Load occupancy map and extract obstacles
-        occ_img = cv2.imread(self.fpath_location_occup_img, cv2.IMREAD_GRAYSCALE)
-        _, self.binary_map = cv2.threshold(occ_img, 128, 255, cv2.THRESH_BINARY)
-        self.obstacle_polygons = self.extract_obstacle_polygons(self.binary_map)
-        self.free_region = self.extract_free_region(self.binary_map)
-        
-        # Data for control points and track
-        self.control_points = []  # List of QPointF
-        self.centerline = None
-        self.left_boundary = None
-        self.right_boundary = None
-        self.boundaries_swapped = False  # False: left=blue, right=yellow
-        
-        # GUI control variables
-        self.mode = "add"  # Modes: "add", "remove", "move"
-        self.selected_point_index = None
-        self.dragging = False
         
         # Create main widget and layout
         self.main_widget = QWidget()
@@ -140,7 +45,7 @@ class FSTrackDraw(QMainWindow):
         self.main_layout = QHBoxLayout(self.main_widget)
         self.main_layout.setContentsMargins(5, 5, 5, 5)
         
-        # Canvas for drawing
+        # Create canvas (handles its own obstacle loading with proper scaling)
         self.canvas = TrackCanvas(self)
         self.canvas.setMinimumSize(600, 600)
         self.main_layout.addWidget(self.canvas, 1)
@@ -212,40 +117,20 @@ class FSTrackDraw(QMainWindow):
         # Add stretch to push everything up
         self.ui_layout.addStretch(1)
         
-        # Initialize canvas
-        self.redraw()
-    
-    def extract_obstacle_polygons(self, binary_map):
-        """Extract obstacle contours from the occupancy map."""
-        contours, _ = cv2.findContours(binary_map, cv2.RETR_CCOMP, cv2.CHAIN_APPROX_SIMPLE)
-        polys = []
-        if contours is not None:
-            for cnt in contours:
-                pts = cnt.squeeze()
-                if pts.ndim == 1:
-                    continue
-                polys.append(pts.tolist())
-        return polys
+        # Data for control points and track
+        self.control_points = []  # List of QPointF
+        self.centerline = None
+        self.left_boundary = None
+        self.right_boundary = None
+        self.boundaries_swapped = False
         
-    def extract_free_region(self, binary_map):
-        """Extract the largest external contour from the occupancy map as the free region."""
-        contours, _ = cv2.findContours(binary_map, cv2.RETR_CCOMP, cv2.CHAIN_APPROX_SIMPLE)
-        if not contours:
-            return None
-        largest = max(contours, key=cv2.contourArea)
-        pts = largest.squeeze()
-        if pts.ndim == 1 or len(pts) < 3:
-            return None
-        if not np.array_equal(pts[0], pts[-1]):
-            pts = np.vstack([pts, pts[0]])
-        try:
-            poly = Polygon(pts)
-            if not poly.is_valid or poly.is_empty:
-                return None
-            return poly
-        except Exception as e:
-            print("Error creating free region polygon:", e)
-            return None
+        # GUI control variables
+        self.mode = "add"  # Modes: "add", "remove", "move"
+        self.selected_point_index = None
+        self.dragging = False
+        
+        # Initialize
+        self.redraw()
     
     def activate_add_mode(self):
         self.mode = "add"
@@ -360,9 +245,7 @@ class FSTrackDraw(QMainWindow):
             self.centerline,
             self.left_boundary,
             self.right_boundary,
-            self.boundaries_swapped,
-            self.obstacle_polygons,
-            self.free_region
+            self.boundaries_swapped
         )
         
         # Update track parameters
@@ -402,8 +285,6 @@ class FSTrackDraw(QMainWindow):
             self.min_radius_label.setText(f"Min Radius: {min_radius:.2f} m")
             self.cone_count_label.setText(f"Cones: Blue: {blue_cones} Yellow: {yellow_cones} Total: {total_cones}")
             
-        self.canvas.update()
-        
     def calculate_track_length(self):
         """Calculate the total length of the track (centerline B-spline)."""
         if len(self.centerline) < 2:
@@ -472,134 +353,3 @@ class FSTrackDraw(QMainWindow):
             return len(right_cones), len(left_cones), len(right_cones) + len(left_cones)
         else:
             return len(left_cones), len(right_cones), len(left_cones) + len(right_cones)
-
-
-class TrackCanvas(QWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.parent = parent
-        self.setMouseTracking(True)
-        
-        # Load the satellite image
-        self.sat_image = QImage(parent.fpath_location_sat_img)
-        self.sat_pixmap = QPixmap.fromImage(self.sat_image)
-        
-        # Drawing elements
-        self.control_points = []
-        self.centerline = None
-        self.left_boundary = None
-        self.right_boundary = None
-        self.boundaries_swapped = False
-        self.obstacle_polygons = []
-        self.free_region = None
-        
-    def update_drawing(self, control_points, centerline, left_boundary, right_boundary, 
-                      boundaries_swapped, obstacle_polygons, free_region):
-        self.control_points = control_points
-        self.centerline = centerline
-        self.left_boundary = left_boundary
-        self.right_boundary = right_boundary
-        self.boundaries_swapped = boundaries_swapped
-        self.obstacle_polygons = obstacle_polygons
-        self.free_region = free_region
-        
-    def paintEvent(self, event):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.Antialiasing)
-        
-        # Draw satellite image
-        painter.drawPixmap(0, 0, self.sat_pixmap.scaled(self.width(), self.height(), Qt.KeepAspectRatio))
-        
-        # Draw obstacles
-        painter.setPen(QPen(QColor(255, 0, 0), 2))
-        for poly in self.obstacle_polygons:
-            qpoly = QPolygonF([QPointF(p[0], p[1]) for p in poly])
-            painter.drawPolygon(qpoly)
-        
-        # Draw safe backoff region
-        try:
-            backoff_val = float(self.parent.backoff_entry.text())
-        except ValueError:
-            backoff_val = self.parent.min_boundary_backoff
-            
-        backoff_px = backoff_val * self.parent.px_per_m
-        if self.free_region is not None:
-            safe_region = self.free_region.buffer(-backoff_px)
-            if safe_region and not safe_region.is_empty and safe_region.exterior is not None:
-                coords = list(safe_region.exterior.coords)
-                if len(coords) >= 4:
-                    qpoly = QPolygonF([QPointF(p[0], p[1]) for p in coords])
-                    pen = QPen(QColor(255, 0, 255), 2)
-                    pen.setStyle(Qt.DashLine)
-                    painter.setPen(pen)
-                    painter.drawPolygon(qpoly)
-        
-        # Draw control points
-        for i, pt in enumerate(self.control_points):
-            if i == 0 and len(self.control_points) > 2:
-                painter.setBrush(QColor(255, 0, 0))  # Red for first point
-            elif i == len(self.control_points) - 1 and len(self.control_points) > 2:
-                painter.setBrush(QColor(0, 0, 255))  # Blue for last point
-            else:
-                painter.setBrush(QColor(0, 255, 0))  # Green for others
-            painter.setPen(QPen(QColor(0, 0, 0), 1))
-            painter.drawEllipse(pt, 5, 5)
-        
-        # Draw centerline
-        if self.centerline and len(self.centerline) > 1:
-            pen = QPen(QColor(0, 255, 0), 2)
-            pen.setStyle(Qt.DashLine)
-            painter.setPen(pen)
-            for i in range(1, len(self.centerline)):
-                painter.drawLine(self.centerline[i-1], self.centerline[i])
-        
-        # Draw boundaries
-        if self.left_boundary and len(self.left_boundary) > 1:
-            left_color = QColor(255, 255, 0) if self.boundaries_swapped else QColor(0, 0, 255)
-            pen = QPen(left_color, 2)
-            painter.setPen(pen)
-            for i in range(1, len(self.left_boundary)):
-                painter.drawLine(self.left_boundary[i-1], self.left_boundary[i])
-        
-        if self.right_boundary and len(self.right_boundary) > 1:
-            right_color = QColor(0, 0, 255) if self.boundaries_swapped else QColor(255, 255, 0)
-            pen = QPen(right_color, 2)
-            painter.setPen(pen)
-            for i in range(1, len(self.right_boundary)):
-                painter.drawLine(self.right_boundary[i-1], self.right_boundary[i])
-        
-        # Draw cones
-        try:
-            cone_spacing = float(self.parent.cone_spacing_entry.text())
-        except ValueError:
-            cone_spacing = self.parent.default_cone_distance
-            
-        if self.left_boundary and len(self.left_boundary) > 1:
-            left_boundary = np.array([[p.x(), p.y()] for p in self.left_boundary])
-            left_cones = sample_cones(left_boundary, cone_spacing, self.parent.px_per_m)
-            
-            left_color = QColor(255, 255, 0) if self.boundaries_swapped else QColor(0, 0, 255)
-            painter.setBrush(left_color)
-            painter.setPen(QPen(QColor(0, 0, 0), 1))
-            for pt in left_cones:
-                painter.drawEllipse(QPointF(pt[0], pt[1]), 3, 3)
-                
-        if self.right_boundary and len(self.right_boundary) > 1:
-            right_boundary = np.array([[p.x(), p.y()] for p in self.right_boundary])
-            right_cones = sample_cones(right_boundary, cone_spacing, self.parent.px_per_m)
-            
-            right_color = QColor(0, 0, 255) if self.boundaries_swapped else QColor(255, 255, 0)
-            painter.setBrush(right_color)
-            painter.setPen(QPen(QColor(0, 0, 0), 1))
-            for pt in right_cones:
-                painter.drawEllipse(QPointF(pt[0], pt[1]), 3, 3)
-        
-    def mousePressEvent(self, event):
-        self.parent.handle_canvas_click(event.pos())
-        
-    def mouseMoveEvent(self, event):
-        if self.parent.dragging:
-            self.parent.handle_canvas_drag(event.pos())
-            
-    def mouseReleaseEvent(self, event):
-        self.parent.handle_canvas_release(event.pos())
